// src/constants/character.enum.ts
export enum CharacterEnum {
  Mentor = 'Mentor',
  Guard = 'Guard',
  Messenger = 'Messenger',
  Ally = 'Ally',
  Enemy = 'Enemy',
  Shadow = 'Shadow',
}

export enum GenderEnum {
  Male = 'Male',
  Female = 'Female',
}

export enum AllyTypeEnum {
  BosomFriend = 'BosomFriend',
  HighWorldAlly = 'HighWorldAlly',
  Animal = 'Animal',
  UndeadAlly = 'UndeadAlly',
  AgileServant = 'AgileServant',
}

export enum MentorTypeEnum {
  Dark = 'Dark',
  Fallen = 'Fallen',
  Permanent = 'Permanent',
  Comic = 'Comic',
  Shaman = 'Shaman',
}

export enum KnowledgeTypeEnum {
  Education = 'Education',
  PresentationGifts = 'PresentationGifts',
  TeacherAdvice = 'TeacherAdvice',
  Motivation = 'Motivation',
  SproutsInformation = 'SproutsInformation',
}